// src/plugins/data_view_management/public/components/edit_index_pattern/scripted_fields_table/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0 and the Server Side Public License, v 1; you may not use this file except
 * in compliance with, at your election, the Elastic License 2.0 or the Server
 * Side Public License, v 1.
 */
import type * as estypes from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
/** @internal **/
export interface ScriptedFieldItem {
  name: string;
  lang: estypes.ScriptLanguage;
  script: string;
  isUserEditable?: boolean;
}